+++ rust-src/bulletproofs/src/lib.rs
@@ -6,4 +6,5 @@
 pub mod inner_product_proof;
 pub mod range_proof;
 pub mod set_membership_proof;
+pub mod set_non_membership_proof;
 pub mod utils;